bullet now goes towards creeps

+++ core/src/com/finalstand/game/sprites/projectiles/Bullet.java
@@ -9,7 +9,7 @@ import com.finalstand.game.FinalStand;
  */
 public class Bullet extends Projectile{
 
-    public Bullet(float x, float y, double angle, int level, float w, float h)
+    public Bullet(float x, float y, float angle, int level, float w, float h)
     {
         super(x, y, angle, level);
         projectileSprite = new Sprite(new Texture("projectiles/singleshot_projectile.png"));
